# portal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AST Institutional Portal API
  version: 1.0.0
  description: |
    Edge API for the institutional portal (AFC-side admission).

    **Canon alignment (AST-CORE-CANON):**
    - Portal does **not** mint, burn, or hold third-party funds.
    - `processId` = `AST-{INST}-{YYYYMMDD}-{suffix}` (not UUID).
    - Institutional valuation is accepted as given (decimal string); portal does not appraise.
    - Document evidence: package hash + qualified signature attestation (full X.509 chain = follow-on).
    - Economic finality only after Core Orchestrator + PoT on NodeChain.

    **Auth (v1):** session after login (`X-Session-Id`).  
    **Target:** Bearer JWT + mutual TLS (documented; not required in local demo).
servers:
  - url: http://localhost:3100
    description: Local portal edge (BFF)
  - url: https://api.arosfinancialcore.com
    description: Production portal edge (when deployed)
...
security:
  - SessionAuth: []
paths:
...
  /v1/public/processes/{processId}:
    get:
      tags: [public]
      operationId: publicGetProcess
      security: []
      summary: Public process lookup (no registration, no key)
      description: Read-only redacted view. No mint. No NodeChain write.
      parameters:
        - $ref: '#/components/parameters/ProcessId'
      responses:
        '200':
          description: Public process view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/public/nodechain/status:
    get:
      tags: [public]
      operationId: publicNodechainStatus
      security: []
      summary: NodeChain journal status (tip, chain, engine)
      description: Read-only proxy to Core. Portal never appends.
      responses:
        '200':
          description: Status
        '503':
          description: Core unavailable
  /v1/public/nodechain/blocks:
    get:
      tags: [public]
      operationId: publicNodechainBlocks
      security: []
      summary: Latest blocks (tip-first blockchain-style feed)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 200
            default: 40
      responses:
        '200':
          description: Blocks list
...
  /v1/public/nodechain/verify:
    get:
      tags: [public]
      operationId: publicNodechainVerify
      security: []
      summary: Full hash-chain verify
      responses:
        '200':
          description: Chain OK
        '409':
          description: Chain broken
  /v1/public/nodechain/records/height/{height}:
    get:
      tags: [public]
      operationId: publicNodechainByHeight
      security: []
      summary: Journal record by height
      parameters:
        - name: height
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Record
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/public/nodechain/records/id/{recordId}:
    get:
      tags: [public]
      operationId: publicNodechainByRecordId
      security: []
      summary: Journal record by recordId
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Record
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/public/nodechain/processes/{processId}:
    get:
      tags: [public]
      operationId: publicNodechainByProcess
      security: []
      summary: Process-scoped journal history
      parameters:
        - $ref: '#/components/parameters/ProcessId'
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 500
      responses:
        '200':
          description: Records list
...
  /v1/processes/{processId}:
    get:
      tags: [processes]
      operationId: getProcess
      summary: Process status (edge + Core merge when available)
      parameters:
        - $ref: '#/components/parameters/ProcessId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Current process state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: Session from POST /v1/auth/login (v1 demo / local)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Target production institution OIDC/JWT
    MutualTLS:
      type: mutualTLS
      description: Target production institution mTLS
  parameters:
    SessionId:
      name: X-Session-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
    ProcessId:
      name: processId
      in: path
      required: true
      schema:
        type: string
        pattern: '^AST-[A-Za-z0-9]+-\d{8}-[A-Za-z0-9]+$'
        example: AST-DEMO-20260720-a1b2c3d4e5f6
  schemas:
...
    EdgeProcessStatus:
      type: string
      enum:
        - accepted
        - awaiting_core
        - documents_pending
        - submitted_to_core
        - rejected
        - duplicate
        - created
        - pot_pending
        - completed
        - failed
        - expired
    AssetType:
      type: string
      enum: [real_estate, bond, investment_package, other]
...
    ProcessResponse:
      type: object
      properties:
        processId:
          type: string
          example: AST-DEMO-20260720-a1b2c3d4e5f6
        claimId:
          type: string
          nullable: true
          description: Claim id (v1 equals processId)
        status:
          $ref: '#/components/schemas/EdgeProcessStatus'
        currentStep:
          type: string
          description: Human/pipeline step label
        institutionId:
          type: string
        valuation:
          type: string
        holderId:
          type: string
        assetType:
          $ref: '#/components/schemas/AssetType'
        documentPackageHash:
          type: string
        idempotencyKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [edge, core]
        message:
          type: string
        core:
          type: object
          additionalProperties: true
        coreError:
          type: object
          additionalProperties: true
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...